# expense-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Expense Management API",
        "description": "Multi-Tenant SaaS Expense Management API — secure, role-based, company-scoped.",
        "contact": {
            "email": "gplanet.tech@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local / Development"
        },
        {
            "url": "https://api.yourdomain.com/api",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List company users (Admin only, paginated, role-filterable)",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Admin",
                                "Manager",
                                "Employee"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated user list (served from Redis cache)"
                    },
                    "403": {
                        "description": "Admin role required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a user in the Admin's company (async welcome email dispatched)",
                "operationId": "a1fce870c566eb5fc0cf3178ec5d2fed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Bob"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "bob@acme.com"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "Admin",
                                            "Manager",
                                            "Employee"
                                        ]
                                    },
                                    "password": {
                                        "description": "Optional — auto-generated if omitted",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created"
                    },
                    "403": {
                        "description": "Admin role required"
                    },
                    "422": {
                        "description": "Duplicate email within company or invalid role"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user's name, email, or role (Admin only)",
                "operationId": "ddfe232db1870b82493bf8bd25596d37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "Admin",
                                            "Manager",
                                            "Employee"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated"
                    },
                    "403": {
                        "description": "Admin role required"
                    },
                    "404": {
                        "description": "User not found or cross-company"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user (Admin only — self-delete returns 422)",
                "operationId": "43dd20becb1f47444e484cbb18ce345c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted"
                    },
                    "403": {
                        "description": "Admin role required"
                    },
                    "404": {
                        "description": "User not found or cross-company"
                    },
                    "422": {
                        "description": "Cannot delete own account"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum token — obtain via /auth/login or /auth/register",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}